Split ansitopng.go into files
