+++ internal/formatter/ansitopng/ansitopng.go
@@ -6,6 +6,7 @@ import (
 	"image/color"
 	"image/png"
 	"io/fs"
+	"log"
 	"path/filepath"
 	"strings"
 
@@ -24,24 +25,7 @@ const (
 	FONT_SIZE   = 13
 )
 
-var fontPaths = map[string]string{
-	"default":  "fonts/DejaVuSansMono.ttf",
-	"Cyrillic": "fonts/DejaVuSansMono.ttf",
-	"Greek":    "fonts/DejaVuSansMono.ttf",
-	"Arabic":   "fonts/DejaVuSansMono.ttf",
-	"Hebrew":   "fonts/DejaVuSansMono.ttf",
-	"Han":      "fonts/wqy-zenhei.ttc",
-	"Hiragana": "fonts/MTLc3m.ttf",
-	"Katakana": "fonts/MTLc3m.ttf",
-	"Hangul":   "fonts/LexiGulim.ttf",
-	"Braille":  "fonts/Symbola_hint.ttf",
-	"Emoji":    "fonts/Symbola_hint.ttf",
-}
-
-var weatherSymbolWidth = map[string]int{
-	// Add wide characters here when needed (e.g. weather icons)
-	// "☀": 2, "🌧": 2,
-}
+var weatherSymbolWidth = map[string]int{}
 
 // RenderANSI renders ANSI text to PNG using strongly-typed PNGOptions.
 func RenderANSI(text string, opts PNGOptions) ([]byte, error) {
@@ -140,66 +124,14 @@ func lineLength(line []te.Cell) int {
 	return 0
 }
 
-// colorFromANSI converts go-te color (from Attr.Fg / Attr.Bg) to real color.Color
-func colorFromANSI(c te.Color, inverse bool) color.Color {
-	if c.Mode == te.ColorDefault || c.Name == "default" {
-		if inverse {
-			return color.Black
-		}
-		return color.RGBA{211, 211, 211, 255} // lightgray
-	}
-
-	// Named color
-	switch c.Name {
-	case "black":
-		return color.Black
-	case "red":
-		return color.RGBA{205, 49, 49, 255}
-	case "green":
-		return color.RGBA{0, 128, 0, 255}
-	case "yellow":
-		return color.RGBA{205, 205, 0, 255}
-	case "blue":
-		return color.RGBA{0, 0, 205, 255}
-	case "magenta":
-		return color.RGBA{205, 0, 205, 255}
-	case "cyan":
-		return color.RGBA{0, 205, 205, 255}
-	case "white":
-		return color.RGBA{229, 229, 229, 255}
-	}
-
-	// 256-color mode (most important for wttr.in)
-	if c.Mode == te.ColorANSI256 {
-		return ansi256ToColor(int(c.Index))
-	}
-
-	return color.RGBA{211, 211, 211, 255}
-}
-
-// ansi256ToColor maps common 256-color values used by wttr.in
-func ansi256ToColor(n int) color.Color {
-	switch {
-	case n == 226 || n == 227 || n == 190 || n == 220: // yellow / sun
-		return color.RGBA{255, 215, 0, 255}
-	case n >= 46 && n <= 82: // green shades
-		return color.RGBA{0, 255, 0, 255}
-	case n >= 118 && n <= 154: // bright green / cyan
-		return color.RGBA{0, 255, 100, 255}
-	case n >= 196 && n <= 208: // red / orange
-		return color.RGBA{255, 60, 60, 255}
-	case n >= 33 && n <= 39: // blue
-		return color.RGBA{100, 100, 255, 255}
-	}
-	return color.RGBA{200, 200, 200, 255}
-}
-
 func loadEmojiLib() (map[string]image.Image, error) {
 	emojilib := make(map[string]image.Image)
 	emojiFS, err := fs.Sub(assets.FS, "share/emoji")
 	if err != nil {
-		return nil, err
+		log.Printf("Warning: could not open emoji assets: %v", err)
+		return emojilib, err
 	}
+
 	err = fs.WalkDir(emojiFS, ".", func(path string, d fs.DirEntry, err error) error {
 		if err != nil || d.IsDir() || !strings.HasSuffix(path, ".png") {
 			return err
@@ -209,15 +141,18 @@ func loadEmojiLib() (map[string]image.Image, error) {
 			return err
 		}
 		defer f.Close()
+
 		img, _, err := image.Decode(f)
 		if err != nil {
 			return err
 		}
+
 		resized := resizeImage(img, CHAR_HEIGHT, CHAR_HEIGHT)
 		char := strings.TrimSuffix(filepath.Base(path), ".png")
 		emojilib[char] = resized
 		return nil
 	})
+
 	return emojilib, err
 }
 
@@ -265,10 +200,8 @@ func genTerm(buf [][]te.Cell, graphemes []string, opts PNGOptions) ([]byte, erro
 	for _, line := range buf {
 		xPos := 0.0
 		for _, cell := range line {
-			// FIXED: Use cell.Attr.Fg and cell.Attr.Bg (correct fields in go-te)
 			fg := colorFromANSI(cell.Attr.Fg, opts.Inverted)
 
-			// Background support
 			if cell.Attr.Bg.Mode != te.ColorDefault && cell.Attr.Bg.Name != "default" {
 				bgCol := colorFromANSI(cell.Attr.Bg, opts.Inverted)
 				dc.SetColor(bgCol)
@@ -286,6 +219,10 @@ func genTerm(buf [][]te.Cell, graphemes []string, opts PNGOptions) ([]byte, erro
 
 			if data != "" && data != " " {
 				cat := scriptCategory([]rune(data)[0])
+
+				// Load and set font right before drawing
+				loadAndSetFont(dc, cat)
+
 				if cat == "Emoji" {
 					if img, ok := emojilib[data]; ok {
 						dc.DrawImage(img, int(xPos), int(yPos))
@@ -344,6 +281,26 @@ func getSymbolWidth(s string) int {
 func scriptCategory(r rune) string {
 	if isEmoji(r) {
 		return "Emoji"
+	}
+	switch {
+	case r >= 0x0400 && r <= 0x04FF:
+		return "Cyrillic"
+	case r >= 0x0370 && r <= 0x03FF:
+		return "Greek"
+	case r >= 0x0600 && r <= 0x06FF:
+		return "Arabic"
+	case r >= 0x0590 && r <= 0x05FF:
+		return "Hebrew"
+	case (r >= 0x4E00 && r <= 0x9FFF) || (r >= 0x3400 && r <= 0x4DBF):
+		return "Han"
+	case r >= 0x3040 && r <= 0x309F:
+		return "Hiragana"
+	case r >= 0x30A0 && r <= 0x30FF:
+		return "Katakana"
+	case r >= 0xAC00 && r <= 0xD7AF:
+		return "Hangul"
+	case r >= 0x2800 && r <= 0x28FF:
+		return "Braille"
 	}
 	return "default"
 }
